WIP db interface implementation
I’m trying to make sure which implementation of which business rules
needs direct access to the database or will be fine in the management
layer. My guess is most business rules need not be in the db controller
but will be fine in the management since the content of the map is more
or less static. My concern for now is the casting of objects to java
datatypes and how to make efficient db access methods. I also chose
jdbc over hibernate since I don’t (as of yet) have found the necessity
for it as I intend to keep things simple for as long as it’s possible.