DBcontroller finished but not tested
I have no idea if my implementation works, this is the first time I am
writing this kind of code after all. But the basic functionality should
be clear now. By now I am also considering to use Hibernate in the
management layer since the access to the database will be more
superficial.